# app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Chess.com Play Style API",
    "version": "1.0.0",
    "description": "API for analyzing and mimicking chess.com player styles"
  },
  "basePath": "/",
  "schemes": ["http", "https"],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register new API key",
        "parameters": [{
          "in": "body",
          "name": "body",
          "required": true,
          "schema": {
            "type": "object",
            "properties": {
              "email": {"type": "string"},
              "password": {"type": "string"}
            }
          }
        }],
        "responses": {
          "201": {"description": "API key generated successfully"},
          "400": {"description": "Invalid input"}
        }
      }
    },
    "/auth/update": {
      "put": {
        "tags": ["Authentication"],
        "summary": "Update API key",
        "parameters": [{
          "in": "body",
          "name": "body",
          "required": true,
          "schema": {
            "type": "object",
            "properties": {
              "email": {"type": "string"},
              "password": {"type": "string"}
            }
          }
        }],
        "responses": {
          "200": {"description": "API key updated successfully"},
          "401": {"description": "Invalid credentials"},
          "400": {"description": "Invalid input"}
        }
      }
    },
    "/auth/delete": {
      "delete": {
        "tags": ["Authentication"],
        "summary": "Delete API key",
        "parameters": [{
          "in": "body",
          "name": "body",
          "required": true,
          "schema": {
            "type": "object",
            "properties": {
              "email": {"type": "string"},
              "password": {"type": "string"}
            }
          }
        }],
        "responses": {
          "200": {"description": "User deleted successfully"},
          "401": {"description": "Invalid credentials"},
          "400": {"description": "Invalid input"}
        }
      }
    },
    "/api/player/profile/{username}": {
      "get": {
        "tags": ["Player"],
        "summary": "Get player profile",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "type": "string"
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {"description": "Player profile retrieved successfully"},
          "401": {"description": "Invalid API key"},
          "400": {"description": "Invalid username"}
        }
      }
    },
    "/api/player/style/{username}": {
      "get": {
        "tags": ["Player"],
        "summary": "Get player style analysis",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "type": "string"
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {"description": "Style analysis retrieved successfully"},
          "401": {"description": "Invalid API key"},
          "400": {"description": "Invalid username"}
        }
      }
    },
    "/api/player/mimic-move/{username}": {
      "post": {
        "tags": ["Player Style Mimicry"],
        "summary": "Generate a chess move in the style of the player",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "type": "string",
            "description": "Chess.com username"
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "fen": {
                  "type": "string",
                  "description": "Chess position in FEN notation"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Move generated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "username": {"type": "string"},
                "original_fen": {"type": "string"},
                "move": {"type": "string"},
                "new_fen": {"type": "string"},
                "reasoning": {"type": "string"}
              }
            }
          },
          "401": {"description": "Invalid API key"},
          "400": {"description": "Invalid input or move generation failed"}
        }
      }
    },
    "/api/player/generate-game/{username}": {
      "post": {
        "tags": ["Player Style Mimicry"],
        "summary": "Generate a complete chess game in the style of the player",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "type": "string",
            "description": "Chess.com username"
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "opening": {
                  "type": "string",
                  "description": "Optional opening to start with (e.g., 'Sicilian Defense')"
                },
                "opponent_style": {
                  "type": "string",
                  "description": "Description of opponent's playing style"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Game generated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "username": {"type": "string"},
                "pgn": {"type": "string"},
                "opponent_style": {"type": "string"},
                "opening": {"type": "string"}
              }
            }
          },
          "401": {"description": "Invalid API key"},
          "400": {"description": "Invalid input or game generation failed"}
        }
      }
    }
  }
}